+++ nrkdownload/config.py
@@ -4,7 +4,7 @@ import os
 DOWNLOAD_DIR = os.getenv('NRKDOWNLOAD_DIR', os.path.expanduser('~/Downloads/nrkdownload'))
 
 # This keeps track of the series we have seen
-KNOWN_SERIES = {}
+# KNOWN_SERIES = {}
 
 # What is the maximum string length in the list output when searching?
 MAX_OUTPUT_STRING_LENGTH = 70
